requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.113.0
h11==0.14.0
idna==3.8
protobuf==5.28.0
psycopg2-binary==2.9.9
py4j==0.10.9.7
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
pyspark==3.5.2
python-dotenv==1.0.1
redis==5.0.8
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
